Extrais un adaptateur pour le tampon d'homologation
...et utilise le nouvel adaptateurPdf.typst pour la génération de PDF.
