+++ src/adaptateurs/adaptateurTamponHomologation.configuration.ts
