+++ src/adaptateurs/adaptateurTamponHomologation.puppeteer.ts
