+++ src/mss.js
@@ -22,6 +22,7 @@ const creeServeur = ({
   referentielV2,
   adaptateurMail,
   adaptateurPdf,
+  adaptateurTamponHomologation,
   adaptateurHorloge,
   adaptateurGestionErreur,
   serviceAnnuaire,
@@ -160,6 +161,7 @@ const creeServeur = ({
       referentielV2,
       adaptateurHorloge,
       adaptateurPdf,
+      adaptateurTamponHomologation,
       adaptateurCsv,
       adaptateurZip,
       adaptateurJournal,
