+++ src/routes/connecte/routesConnecteApi.js
@@ -57,6 +57,7 @@ const routesConnecteApi = ({
   referentielV2,
   adaptateurHorloge,
   adaptateurPdf,
+  adaptateurTamponHomologation,
   adaptateurCsv,
   adaptateurZip,
   adaptateurJournal,
@@ -333,6 +334,7 @@ const routesConnecteApi = ({
       referentielV2,
       adaptateurHorloge,
       adaptateurPdf,
+      adaptateurTamponHomologation,
       adaptateurZip,
       adaptateurJournal,
       adaptateurEnvironnement,
